+++ src/php/Gdbots/Schemas/Iam/Mixin/CreateRole/CreateRole.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\CreateRole;
+
+use Gdbots\Schemas\Ncr\Mixin\CreateNode\CreateNodeV1;
+
+interface CreateRole extends CreateNodeV1
+{
+}