+++ src/php/Gdbots/Schemas/Iam/Mixin/DeleteRole/DeleteRole.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\DeleteRole;
+
+use Gdbots\Schemas\Ncr\Mixin\DeleteNode\DeleteNodeV1;
+
+interface DeleteRole extends DeleteNodeV1
+{
+}