+++ src/php/Gdbots/Schemas/Iam/Mixin/DeleteRole/DeleteRoleV1Mixin.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\DeleteRole;
+
+use Gdbots\Pbj\AbstractMixin;
+use Gdbots\Pbj\FieldBuilder as Fb;
+use Gdbots\Pbj\SchemaId;
+use Gdbots\Pbj\Type as T;
+
+final class DeleteRoleV1Mixin extends AbstractMixin
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getId()
+    {
+        return SchemaId::fromString('pbj:gdbots:iam:mixin:delete-role:1-0-0');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getFields()
+    {
+        return [];
+    }
+}