+++ src/php/Gdbots/Schemas/Iam/Mixin/GetRoleBatchRequest/GetRoleBatchRequestV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\GetRoleBatchRequest;
+
+interface GetRoleBatchRequestV1 extends GetRoleBatchRequest
+{
+}