+++ src/php/Gdbots/Schemas/Iam/Mixin/GetRoleBatchResponse/GetRoleBatchResponseV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\GetRoleBatchResponse;
+
+interface GetRoleBatchResponseV1 extends GetRoleBatchResponse
+{
+}