+++ src/php/Gdbots/Schemas/Iam/Mixin/GetRoleHistoryRequest/GetRoleHistoryRequest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\GetRoleHistoryRequest;
+
+use Gdbots\Schemas\Pbjx\Mixin\GetEventsRequest\GetEventsRequestV1;
+
+interface GetRoleHistoryRequest extends GetEventsRequestV1
+{
+}