+++ src/php/Gdbots/Schemas/Iam/Mixin/GetRoleHistoryRequest/GetRoleHistoryRequestV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\GetRoleHistoryRequest;
+
+interface GetRoleHistoryRequestV1 extends GetRoleHistoryRequest
+{
+}