+++ src/php/Gdbots/Schemas/Iam/Mixin/GetRoleHistoryResponse/GetRoleHistoryResponseV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\GetRoleHistoryResponse;
+
+interface GetRoleHistoryResponseV1 extends GetRoleHistoryResponse
+{
+}