+++ src/php/Gdbots/Schemas/Iam/Mixin/GetRoleResponse/GetRoleResponseV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\GetRoleResponse;
+
+interface GetRoleResponseV1 extends GetRoleResponse
+{
+}