+++ src/php/Gdbots/Schemas/Iam/Mixin/GetRoleResponse/GetRoleResponseV1Mixin.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\GetRoleResponse;
+
+use Gdbots\Pbj\AbstractMixin;
+use Gdbots\Pbj\FieldBuilder as Fb;
+use Gdbots\Pbj\SchemaId;
+use Gdbots\Pbj\Type as T;
+
+final class GetRoleResponseV1Mixin extends AbstractMixin
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getId()
+    {
+        return SchemaId::fromString('pbj:gdbots:iam:mixin:get-role-response:1-0-0');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getFields()
+    {
+        return [
+            Fb::create('node', T\MessageType::create())
+                ->className('Gdbots\Schemas\Iam\Mixin\Role\Role')
+                ->build()
+        ];
+    }
+}