+++ src/php/Gdbots/Schemas/Iam/Mixin/ListAllRolesRequest/ListAllRolesRequest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\ListAllRolesRequest;
+
+use Gdbots\Schemas\Pbjx\Mixin\Request\RequestV1;
+
+interface ListAllRolesRequest extends RequestV1
+{
+}