+++ src/php/Gdbots/Schemas/Iam/Mixin/ListAllRolesResponse/ListAllRolesResponseV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\ListAllRolesResponse;
+
+interface ListAllRolesResponseV1 extends ListAllRolesResponse
+{
+}