+++ src/php/Gdbots/Schemas/Iam/Mixin/RoleDeleted/RoleDeletedV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\RoleDeleted;
+
+interface RoleDeletedV1 extends RoleDeleted
+{
+}