+++ src/php/Gdbots/Schemas/Iam/Mixin/RoleUpdated/RoleUpdated.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\RoleUpdated;
+
+use Gdbots\Schemas\Ncr\Mixin\NodeUpdated\NodeUpdatedV1;
+
+interface RoleUpdated extends NodeUpdatedV1
+{
+}