+++ src/php/Gdbots/Schemas/Iam/Mixin/RoleUpdated/RoleUpdatedV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\RoleUpdated;
+
+interface RoleUpdatedV1 extends RoleUpdated
+{
+}