+++ src/php/Gdbots/Schemas/Iam/Mixin/UpdateRole/UpdateRoleV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\UpdateRole;
+
+interface UpdateRoleV1 extends UpdateRole
+{
+}