+++ src/php/Gdbots/Schemas/Iam/Mixin/UpdateRole/UpdateRoleV1Trait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Gdbots\Schemas\Iam\Mixin\UpdateRole;
+
+use Gdbots\Pbj\Schema;
+
+/**
+ * @method static Schema schema
+ * @method mixed get($fieldName, $default = null)
+ */
+trait UpdateRoleV1Trait
+{
+}